fix: zunderscore/chocolatey-packages#4
This change adds a test that will check if the font file to be
uninstalled actually exists in the system.
If the file does not exist in the system, the uninstaller skips the step
of trying to uninstall that font file and giving a non-zero error
because the file does not exist in the system.

`$windowsFonts` variable would ideally point to `C:\Windows\Fonts`.

`$env:SystemRoot` ideally points to `C:\Windows`.